refactor: 컨벤션에 따른 서비스단 추상화 리팩토링
